Slight Tower Rework
Towers now fire bullets based on the animation, and fire rate just affects the animation speed. We only have an animation for the machine gun and the tri-cannon right now, so there's pretty much a standardized firing speed now. The bug with the towers firing before turning is fixed now as well.