Tower System Update
Towers now have upgrade paths, and all data is in the scriptable objects. Not sure if we want to be able to upgrade barriers.

Right now, only the tri cannon and machine gun have any upgrades since they're the only ones with hotkeys.

You also cannot place towers on top of one another now.

The build manager keeps track of the positions of all the barriers and towers in an array, and there's a get function for the array. When saving, maybe find a game object with each position and store that?