Final additions for the sprint
WARNING - MUSIC JUMPSCARE. The music Zach made now plays during a round. I have jumpscared myself a couple times even when I knew it was coming, so be warned.

Cooldown indicators now appear in the corner during a round and show when an ability is no longer on cooldown. I was originally thinking of letting it be possible to click on them, but that just started problems with the sword swing animation.

There is now a circular health bar for the centerpiece in the top right corner. It starts out as green but slowly becomes more red as the health decreases.

The barrel explosion is now in the game and accurately represents the radius of the explosion (though the effect is instant and the animation is slightly delayed).