Level/round scaling, passive abilities, and critical hits
Regular attacks show up as grey damage numbers, crits show up as red and have double damage.

Passive abilities are now in the game. Right now they're just buffs, idk if we want to do anything with the passives that can't be done with a buff.

Enemy health scales linearly with the round set (round / 10), and the player's damage scales exponentially with their level.