Upgrade Menu + Hover text
Clicking on a tower will bring up a menu where you can choose to spend scrap to upgrade the tower or destroy it to get half your total scrap investment back.

There is also hover text that shows up when you hover over the upgrade options or the banner. I want to implement this in more places later (particularly the character select screen and the cooldown indicators)