Various UI work
You can now click on barriers to bring up the same menu as a tower. A fishing net will just have the destroy button, but the barricade will have an option to drag a health bar to repair it. The same will appear when you click on the centerpiece. The more damage it has taken, the more expensive it is to repair. I'll probably make a change to that at some point to make it keep track of how many times you've repaired it and make it that much more expensive. I didn't end up adding a priority changer, but I'll probably add that tomorrow.

Also I changed the enemy and player scaling to be more even.