Slight enemy rework
I made it so that every enemy has the ability to use attacks, you just have to assign them in the inspector. This includes the projectile/buff attacks, but those still need more work before they're fully functional. As of the last time I worked on them, the buffs mostly work, but there's no visual effect, and the bullets behave very strangely.