Achievements, Forgetfulness rework, balancing, etc
Basic achievement system in place, all that's missing is the script that actually tracks the things needed to unlock them.

Forgetting an enemy now makes them invisible and intangible to you, not the towers

Small balance changes, like increasing the price on the scrap turret, lowering the power of one of the torpedo launcher upgrades, raising the price of tri cannon vortex engine upgrade, etc.

Added functionality for the centerpiece to have a shield, but nothing activates it yet. Hasn't been tested yet, but it should work in  theory.

Performance issues probably aren't fixed, but they're improved.

Temp towers (Old Man's old friend) don't count down time in between rounds, and you can click on them to have the option to dismiss them. I'm thinking of having an indicator of how much time is left for them too.

Etc.