Bug Fixes + Small Reworks
I think I got the big one.  It was a missing reference in the Shader manager. However, some people were saying they got stuck on round 5 or 7, and I haven't been able to replicate that. Maybe something else I changed fixed that.

Reworked the Centerpiece repairs slightly to make the entire bar the amount that you heal instead of the total health of the barrier.

Reworked the Melee attacks of the pirate. I think the reason it was so inconsistent was that the overlapping animations in the blend tree were making the wrong hitbox come out, so I made it calculate what hitbox to use based on the mouse instead of the animation